# Sort filepatterns for faster searching