- [x] This comment contains a description of changes (with reason)

Search patterns are already divided into categories to ensure the simplest to match patterns are run first. 

I noticed that the "*.json" files run particularly long. It turns out that that `afterqc` has 10_000 `num_lines`. Sequali, fastp, and cutadapt have 10, 50 and 100 `num_lines` respectively. It seems only fair that these are run first.

Speed as reported by multiqc:
before on the test-data dir: `9.90s: Searching files`
after on the test-data dir: `9.47s: Searching files`
before on a dir with 800 sequali only reports: `3.48s: Searching files`
after on a dir with 800 sequali only reports: `0.38s: Searching files`

As the sequali author, I would be very happy if this gets merged ;-). Also on very slow filesystems such as NFS the differences would be more significant. 

I know a posted a lot of PRs today, but it seemed better to have these all separately evaluated. This is the last of that series.