@vladsavelyev 2024-04-29
> I agree with all your comments, but since it is merged already, I will just let the code rest, if that's okay. Thanks for merging!

Sorry if I merged too quickly, I'm so excited about the optimizations so want all of them in `main` as fast as possible 😄 

If you find anything suboptimal or want to add or refactor, feel free to create even trivial PRs and I'll merge quickly, @rhpvorderman