@vladsavelyev 2024-04-29
> Those were your comments after all

I meant to merge the proposed change before merging the PR 😓  I'll create another one https://github.com/MultiQC/MultiQC/pull/2512